% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_alpha.R
\name{c_alpha}
\alias{c_alpha}
\title{Cronbach's Alpha}
\usage{
c_alpha(m)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}
}
\value{
Cronbach's alpha coefficient between 0 and 1
}
\description{
Computes Cronbach's Alpha
}
