% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Confidence Interval}
\usage{
conf_int(m)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}
}
\value{
Confidence interval for the computed coefficient
}
\description{
Computes confidence interval for Cronbach's Alpha Coefficient
}
