% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_itemtotcor.R
\name{corr_itemtotcor}
\alias{corr_itemtotcor}
\title{Corrected item-total correlations}
\usage{
corr_itemtotcor(m)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}
}
\value{
corrected item-total correlations between 0 and 1
}
\description{
Computes corrected item-total correlations
}
