% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL6.R
\name{GL6}
\alias{GL6}
\title{Guttman's Lambda 6}
\usage{
GL6(m)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}
}
\value{
Guttman's Lambda 6 between 0 and 1
}
\description{
Computes Guttman's Lambda 6
}
