% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_dist.R
\name{ir_dist}
\alias{ir_dist}
\title{Item Response Distribution Plots}
\usage{
ir_dist(m, item)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}

\item{item}{Item label or Column label of the selected item}
}
\value{
Item Response Distribution Plot
}
\description{
Displays item response distribution for the selected item
}
