% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_satisfaction}
\alias{life_satisfaction}
\title{Responses to the life satisfaction survey}
\format{numeric dataframe of responses}
\source{
\url{http://spss.allenandunwin.com.s3-website-ap-southeast-2.amazonaws.com/Files/survey.zip}
}
\usage{
life_satisfaction
}
\description{
A sample dataset containing responses of 436 individuals to a 5-item Life Satisfaction Survey to test the fuctions of the package
}
\keyword{datasets}
