% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_itemtotcor.R
\name{raw_itemtotcor}
\alias{raw_itemtotcor}
\title{Raw item-total correlations}
\usage{
raw_itemtotcor(m)
}
\arguments{
\item{m}{Dataframe of Likert type responses to the scale items}
}
\value{
Raw item-total correlations between 0 and 1
}
\description{
Computes raw item-total correlations
}
